+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 
 - Added `basic-fair-price.py` script to calculate fair values for `tutorial.py`.
 - Sea shells collected: **+2,247**
+
+## [08-04-2025]
+
+- Added `bong.py` which trades all 3 commodities (merge of `kelply.py` and `resin_ink_merge.py`)
+- Kelp: VWAP mid-price then MM
+- Resin: eyeball fair price then MM
+- Ink: Momentum trading (issue with -50 position not recovering)
+- Resin and Kelp perform well, Squid Ink needs work